performance improvements
-ServerWorkers now yield processor time when no packets are available
-Extracted BlockingQueue to it's own class
-Increased javadoc coverage
