+++ src/network/ServerWorker.java
@@ -1,10 +1,13 @@
 package network;
 
+import util.BlockingQueue;
+
 import java.io.DataInputStream;
 import java.io.DataOutputStream;
 import java.io.IOException;
 import java.net.Socket;
-import java.util.Vector;
+
+import static java.lang.Thread.yield;
 
 /**
  * A threaded worker process to handle communications to and from a single client
@@ -19,7 +22,7 @@ class ServerWorker implements Runnable{
     /**
      *
      * @param socket The socket which this worker should transmit and recieve from
-     * @param msgs
+     * @param msgs The blocking queue to deliver messages to
      * @throws IOException If the the socket is unable to produce input and/or output streams
      */
     ServerWorker(Socket socket, BlockingQueue msgs) throws IOException {
@@ -31,16 +34,24 @@ class ServerWorker implements Runnable{
 
     @Override
     public void run() {
+        //TODO determine if a client has disconnected and terminate thread
         try{
 
             isRunning = true;
             while (isRunning){
+
+                //readUTF() blocks until success, so we must check before calling it to avoid waiting if a packet isnt ready
                 if (dataInputStream.available() > 0){
-                    //TODO deliver the message instead of discarding
                     String msg = dataInputStream.readUTF();
-                    System.out.println("Incoming Message from " + socket.getInetAddress() + ":" + socket.getPort() + " > " +  msg);  //TODO also replace print statements with logging framework
+
+                    //TODO replace print statements with logging framework
+                    System.out.println("Incoming Message from " + socket.getInetAddress() + ":" + socket.getPort() + " > " +  msg);
+
+                    //add the recieved message to the queue for the server to broadcast later
                     msgs.add(msg);
                 }
+
+                yield();
             }
 
 
@@ -65,7 +76,7 @@ class ServerWorker implements Runnable{
      * @param msg The string to be transmitted. It does not need to be null terminated.
      * @throws IOException if the dataOutputStream fails to send
      */
-    public void sendUTF(String msg)throws IOException{
+    void sendUTF(String msg)throws IOException{
         dataOutputStream.writeUTF(msg);
         dataOutputStream.flush();
     }
